micro-infra-spring-base config rewrite groovy to java
